# Improve spectrum analyzer performance by caching most used computations