--- a/plugins/SpectrumAnalyzer/SaSpectrumView.cpp
+++ b/plugins/SpectrumAnalyzer/SaSpectrumView.cpp
@@ -25,7 +25,7 @@
  */
 
 #include "SaSpectrumView.h"
-#include <iostream>
+
 #include <algorithm>
 #include <cmath>
 #include <QMouseEvent>