As previously discussed on Discord, my first attempt at improving the spectrum view improved the performance by around 4 % and needed 16 extra lines, while making the code less readable. So it did not seem worthwhile to merge that.

But. People say there is no such thing as a free lunch, but after thinking about it a bit more, I was able to bump that to 22 lines and a 100 % performance improvement. If that's not close to a free lunch, I don't know what is.

Compared to the first attempt, I now moved the caching closer to the point of use and extended it to a full LUT of complete results, instead of just pre-computed parts of a formula. Before the optimization, the `QPainter` path building took 9.5 ms out of 13.5 ms for the whole redraw (stereo and peak hold enabled, default TripleOsc tone playing, default analyzer window size on a 4K screen). After the update, path build takes only 1.8 ms and and the full redraw is 6.5 ms (there is also an extra 0.7 ms difference, not sure where that came from). That is, the maximum FPS limit increased from 74 to 154, leaving much more headroom for other plugins and GUI elements.