@he29-net 2021-05-08
OK, done. The Bass range problem was really unrelated, so maybe the alpha just runs better due to better compile-time optimization or the lack of profiling probes. (The problem was that the Bass range is the most "zoomed in" in respect to individual bins, so when considering the pixel range to draw, the total image width at that zoom level turns out to be around 160 000 pixels. Even though only like 2000 pixels out of it will be drawn, each pixel was tested to see if it belongs to the range useful for drawing. So all that was needed to fix this was to move the test directly to the for cycle init / end condition, to make sure it does not look outside of the useful range at all.)