@PhysSong 2021-05-20
Looks good now. BTW, can you guarantee the cache miss condition never changes, or not?