@he29-net 2021-05-20
> Looks good now. BTW, can you guarantee the cache miss condition never changes, or not?

I don't think there is any way to guarantee it, but I can't currently think of any future improvement that could affect it. It currently triggers on all inputs of `SaProcessor::freqToXPixel()`, so it should be pretty safe. Just to be sure, I added a comment to that function, explaining that the results may be cached and any new inputs added to the function in the future must also trigger cache update.

Looking deeper, there is also `SaProcessor::binToFreq()`, which could in theory affect the result if sample rate / Nyquist frequency changes. But since LMMS does not support runtime changes of sample rate, this should not be a problem.

--- OK, wait: some time ago I was able to observe sample rate changes visible in the analyzer when exporting at a higher sample rate than LMMS currently uses. First I wasn't able to reproduce it (i.e. the "auto" X range did not adjust to the higher Nyquist limit during export), but I just found out the problem is that OGG export silently ignores the selected sample rate setting and exports at 48 kHz (it would make sense for Opus, which uses fixed 48 kHz, but I thought OGG in LMMS refers to OGG Vorbis.. either way a bit confusing).

When I do a high sample rate export in FLAC, I can still see the change in X range. But because of the cache, the actual displayed spectrum matches the old range instead. I'm not sure if I should add sample rate as another trigger to rebuild the cache. It would be "more correct" and future-proof, but it would be also one more condition that needs to be checked every time and basically never changes. Still, I'm inclined to do it, considering it may save some trouble in the future and compared to the cache savings the impact of one more check is insignificant.