@he29-net 2021-05-21
Sample rate check added. Should be pretty bullet-proof now. I think.